Issue #3: Made sure previous functionality still works, fixed some bugs.
One thing I fixed that wasn't a bug with the new code is that the conditions
across clones were not independent. Since it was a list (and previously a
dict), when the creature's __dict__ was copied only a reference was copied.
I fixed it to copy a full slice of the conditions list.